modified:   ggd/src/components/shop/BproductUpdata.jsx 	modified:   ggd/src/components/shop/BproductWirte.jsx 	new file:   ggd/src/components/shop/Option.jsx 	modified:   ggdserver/src/main/java/com/icia/ggdserver/service/BproductService.java 	renamed:    ggdserver/src/main/webapp/productupload/1730790574620.png -> ggdserver/src/main/webapp/productupload/1730795558698.png

+++ ggdserver/src/main/java/com/icia/ggdserver/service/BproductService.java
@@ -227,11 +227,17 @@ public class BproductService {
 
         try {
 
-            List<Long> deleimg = new ArrayList<>();
-            deleimg.add( bproductTbl.getBpnum());
-            List<String> deleimgList = bpfRepo.selectBproductfilesysnames(deleimg);
-            deleteFileList(deleimgList,session);
-            bpfRepo.deleteAllByBproductfilenums(deleimg);
+            if(bresult != null) {
+                List<Long> deleimg = new ArrayList<>();
+                deleimg.add( bproductTbl.getBpnum());
+                List<String> deleimgList = bpfRepo.selectBproductfilesysnames(deleimg);
+                deleteFileList(deleimgList,session);
+                bpfRepo.deleteAllByBproductfilenums(deleimg);
+            }
+
+
+
+
 
             if (bupdatefiles != null && !bupdatefiles.isEmpty()){
                 buploadFile(bupdatefiles, session, bproductTbl);
